arch/arm/mmu: unify all uint32_t & uintptr_t
Only keep cp15 API use uint32_t, other mmu all use uintptr_t

+++ arch/arm/src/am335x/am335x_boot.c
@@ -207,7 +207,7 @@ static void am335x_vectorpermissions(uint32_t mmuflags)
    * table
    */
 
-  uint32_t pte = mmu_l2_getentry(PG_L2_VECT_VADDR, 0);
+  uintptr_t pte = mmu_l2_getentry(PG_L2_VECT_VADDR, 0);
 
   /* Mask out the old MMU flags from the page table entry.
    *
